src/conda_requirement.txt
 This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.3.0=pypi_0
adtk=0.6.2=pypi_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
argon2-cffi=20.1.0=py39h2bbff1b_1
asttokens=2.0.8=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bokeh=2.4.3=pyhd8ed1ab_3
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39hb82d6ee_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.9.24=h5b45459_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=pyhd8ed1ab_0
cesium=0.10.1=pypi_0
cffi=1.15.1=py39h0878f49_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py39hcbf5309_0
cloudpickle=2.2.0=pyhd8ed1ab_0
cmdstan=2.29.2=hc4b6281_2
cmdstanpy=1.0.7=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
contourpy=1.0.5=py39h1f6ef14_0
convertdate=2.4.0=pyhd8ed1ab_0
cryptography=37.0.1=py39h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cvxopt=1.3.0=py39h474676c_2
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.32=py39h415ef7b_0
cytoolz=0.12.0=py39hb82d6ee_0
dask=2022.9.2=pypi_0
dask-core=2022.2.1=pyhd3eb1b0_0
debugpy=1.6.3=py39h415ef7b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2022.9.2=pypi_0
dsdp=5.8=h6e01ec9_1203
entrypoints=0.4=pyhd8ed1ab_0
ephem=4.1.3=py39ha55989b_4
executing=1.1.0=pyhd8ed1ab_0
fftw=3.3.10=nompi_h38027f0_105
fonttools=4.37.4=py39ha55989b_0
freetype=2.12.1=h546665d_0
fsspec=2022.8.2=pyhd8ed1ab_0
gatspy=0.3=pypi_0
glpk=5.0=h8ffe710_0
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.49.1=pypi_0
gsl=2.7=hdfb1a43_0
heapdict=1.0.1=py_0
hijri-converter=2.2.4=pyhd8ed1ab_0
holidays=0.16=pyhd8ed1ab_0
hydra-core=1.2.0=pyhd8ed1ab_0
icu=68.2=h0e60522_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.0.0=pypi_0
importlib_resources=5.10.0=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.16.0=pyh025b116_0
ipython=8.5.0=pyh08f2357_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jsonschema=4.4.0=py39haa95532_0
jupyter_client=7.3.5=pyhd8ed1ab_0
jupyter_core=4.11.1=py39hcbf5309_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=py39h2e07f2f_0
korean_lunar_calendar=0.3.1=pyhd8ed1ab_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=16_win64_mkl
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=16_win64_mkl
libpng=1.6.38=h19919ed_0
libprotobuf=3.20.1=h12be248_4
libsodium=1.0.18=h8d14728_1
libsqlite=3.39.4=hcfcfb64_0
libtiff=4.4.0=h8e97e67_4
libuv=1.44.2=h8ffe710_0
libwebp=1.2.4=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libzlib=1.2.12=hcfcfb64_4
llvmlite=0.39.1=pypi_0
locket=1.0.0=pyhd8ed1ab_0
lunarcalendar=0.0.9=py_0
lxml=4.9.1=pypi_0
m2-bash=4.3.042=5
m2-coreutils=8.25=102
m2-filesystem=2016.04=4
m2-gcc-libs=5.3.0=4
m2-gmp=6.1.0=3
m2-libiconv=1.14=3
m2-libintl=0.19.7=4
m2-msys2-runtime=2.5.0.17080.65c939c=3
m2-sed=4.2.2=3
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-toolchain_win-64=2.4.0=0
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39hb82d6ee_1
matplotlib=3.6.0=py39hcbf5309_0
matplotlib-base=3.6.0=py39haf65ace_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
matrixprofile=1.1.10=py39h080aedc_0
mistune=0.8.4=py39h2bbff1b_1000
mkl=2022.1.0=h6a75c08_874
msgpack-python=1.0.4=py39h2e07f2f_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.4.4=py39haa95532_0
nbformat=5.3.0=py39haa95532_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
neuralprophet=0.3.2=py39haa95532_0
notebook=6.4.11=py39haa95532_0
numba=0.56.2=py39h99ae161_1
numpy=1.23.3=py39h9061af7_0
oauthlib=3.2.1=pypi_0
omegaconf=2.2.3=py39hcbf5309_0
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.0=py39h2ba5b7c_0
pandas-datareader=0.10.0=pypi_0
pandas-montecarlo=0.0.2=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.3.0=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.2.0=py39hdc2b20a_1
pip=22.2.2=pyhd8ed1ab_0
pmdarima=2.0.1=py39hb82d6ee_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.31=pyha770c72_0
prophet=1.1.1=py39hcbf5309_1
protobuf=3.11.2=pypi_0
psutil=5.9.2=py39ha55989b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pykalman=0.9.5=pypi_0
pymeeus=0.5.10=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.12.3=py39hb0d2dfa_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=pyh0701188_6
python=3.9.13=h9a09f29_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.4=pyhd8ed1ab_0
pywin32=303=py39hb82d6ee_0
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=py39hb82d6ee_4
pyzmq=24.0.1=py39hea35a22_0
qt=5.12.9=h5909a2a_4
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.1.2=py39hfd4428b_0
scipy=1.9.1=py39h316f440_0
seaborn=0.12.0=hd8ed1ab_0
seaborn-base=0.12.0=pyhd8ed1ab_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=59.8.0=pypi_0
setuptools-git=1.2=py_1
six=1.16.0=pyh6c4a22f_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.4=hcfcfb64_0
stack_data=0.5.1=pyhd8ed1ab_0
statsmodels=0.13.2=py39h5d4886f_0
stumpy=1.11.1=pyhd8ed1ab_0
tabulate=0.8.10=pypi_0
tbb=2021.6.0=h91493d7_0
tblib=1.7.0=pyhd8ed1ab_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.1=py39haa95532_0
testpath=0.6.0=py39haa95532_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
toolz=0.12.0=pyhd8ed1ab_0
torch-lr-finder=0.2.1=pyhd3eb1b0_0
torchaudio=0.12.1=py39_cu113
torchvision=0.13.1=py39_cu113
tornado=6.1=pypi_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.4.0=pyhd8ed1ab_0
tsfresh=0.19.0=pyhd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022d=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
ujson=5.5.0=py39h99910a6_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=hac3ee0b_8
vs2015_runtime=14.29.30139=h890b9b1_8
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py39haa95532_1
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.2=py39haa95532_0
win_inet_pton=1.1.0=py39hcbf5309_4
winpty=0.4.3=4
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zict=2.2.0=pyhd8ed1ab_0
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.12=hcfcfb64_4
zstd=1.5.2=h7755175_4
